Refactor: runtime

+++ src/libs/libs/dijkstra.h
@@ -3,6 +3,6 @@
 
 #include "graph.h"
 
-long double *dijkstra_algorithm(Graph *graph, int src);
+double *dijkstra_algorithm(Graph *graph, int src);
 
 #endif // _DIJKSTRA_H_
